requirements.txt
psycopg2
python-dotenv
sqlalchemy
eralchemy[graphviz]